Splitting messaging module into 2 files
